+++ libs/minifmod/src/instrument.h
@@ -23,7 +23,7 @@ struct Instrument final
 
 	[[nodiscard]] const Sample& getSample(int note) const noexcept
 	{
-		const uint8_t note_sample = sample_header.note_sample_number[note];
+		const int note_sample = sample_header.note_sample_number[note];
 		assert(note_sample < 16);
 		return sample[note_sample];
 	}
